DB Full Functional

+++ Proyecto_final/Assets/Scripts/Counters/UIFinalTime.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 public class UIFinalTime : MonoBehaviour
 {
     [SerializeField] private TextMeshProUGUI textFinalTime;
-    float finalTime;
+    double finalTime;
     private void Update()
     {
